fix: Notification badge does not appear until notification popover opened
